Add bypass-ops feature to builder (#6012)
### Ticket
Closes [#6011](https://github.com/tenstorrent/tt-mlir/issues/6011)

### Problem description
Builder has no mechanism for bypassing an op that is malfunctioning in
runtime. This feature has been requested

### What's changed
Added list of op locations `self._bypass_ops` to builder, and a function
`builder.bypass(operand)` to add to that list.
Added structure in post op callback to update device tensors with golden
tensors when location matches.

### Checklist
- [ ] New/Existing tests provide coverage for changes